Delete struct Discrete_signal, upd dsp.c funcs; Upd types in std lib.

+++ Efi_libs/std.c
@@ -14,7 +14,7 @@ int64_t randint(int64_t min, int64_t max)
 }
 
 
-void print_array(uint16_t * array, uint64_t len)
+void print_array(uint16_t * array, uint32_t len)
 {
 	for (uint32_t i = 0; i < len; i++)
 		printf("%d ", array[i]);
@@ -22,9 +22,9 @@ void print_array(uint16_t * array, uint64_t len)
 }
 
 
-uint64_t max_index(int16_t * array, uint64_t len)
+uint32_t max_index(int16_t * array, uint32_t len)
 {
-	uint64_t index = 0, max_index = 0;
+	uint32_t index = 0, max_index = 0;
 
 	while (++index < len)
 		if (array[index] > array[max_index])
